+++ src/core/internal/delegatedAccount.test.ts
@@ -1,13 +1,12 @@
 import { Secp256k1, Value } from 'ox'
 import { privateKeyToAccount } from 'viem/accounts'
-import { getBalance, sendTransaction } from 'viem/actions'
-import { signAuthorization } from 'viem/experimental'
+import { getBalance } from 'viem/actions'
 import { describe, expect, test } from 'vitest'
 
 import { getAccount } from '../../../test/src/account.js'
 import { client, delegation } from '../../../test/src/porto.js'
 import * as Call from './call.js'
-import * as Delegation from './delegation.js'
+import * as DelegatedAccount from './delegatedAccount.js'
 import * as Key from './key.js'
 
 describe('execute', () => {
@@ -19,18 +18,18 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
             key,
           }),
         ],
-        delegate: true,
+        initialize: true,
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -41,26 +40,21 @@ describe('execute', () => {
     })
 
     test('delegated: true, key: owner, keysToAuthorize: [P256], executor: JSON-RPC', async () => {
-      const { account, privateKey } = await getAccount(client, {
+      const { account } = await getAccount(client, {
         delegation,
       })
 
-      const authorization = await signAuthorization(client, {
-        account: privateKeyToAccount(privateKey),
-        contractAddress: delegation,
-        delegate: true,
-      })
-      await sendTransaction(client, {
-        authorizationList: [authorization],
-        account: null,
-        to: account.address,
+      await DelegatedAccount.execute(client, {
+        account,
+        calls: [],
+        initialize: true,
       })
 
       const key = Key.createP256({
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -70,7 +64,7 @@ describe('execute', () => {
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -89,19 +83,19 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
             key,
           }),
         ],
-        delegate: true,
+        initialize: true,
         executor: privateKeyToAccount(privateKey),
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -116,22 +110,17 @@ describe('execute', () => {
         delegation,
       })
 
-      const authorization = await signAuthorization(client, {
-        account: privateKeyToAccount(privateKey),
-        contractAddress: delegation,
-        delegate: true,
-      })
-      await sendTransaction(client, {
-        authorizationList: [authorization],
-        account: null,
-        to: account.address,
+      await DelegatedAccount.execute(client, {
+        account,
+        calls: [],
+        initialize: true,
       })
 
       const key = Key.createP256({
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -142,7 +131,7 @@ describe('execute', () => {
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -159,7 +148,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.setCanExecute(),
@@ -167,14 +156,14 @@ describe('execute', () => {
             key,
           }),
         ],
-        delegate: true,
+        initialize: true,
       })
 
       const nextKey = Key.createP256({
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -185,7 +174,7 @@ describe('execute', () => {
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 1,
         }),
@@ -202,7 +191,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.setCanExecute(),
@@ -210,14 +199,14 @@ describe('execute', () => {
             key,
           }),
         ],
-        delegate: true,
+        initialize: true,
       })
 
       const nextKey = await Key.createWebCryptoP256({
         role: 'admin',
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -228,7 +217,7 @@ describe('execute', () => {
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 1,
         }),
@@ -245,10 +234,10 @@ describe('execute', () => {
         delegation,
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [],
-        delegate: true,
+        initialize: true,
       })
 
       const alice = privateKeyToAccount(Secp256k1.randomPrivateKey())
@@ -263,7 +252,7 @@ describe('execute', () => {
       expect(balances_before[1]).toEqual(Value.fromEther('0'))
       expect(balances_before[2]).toEqual(Value.fromEther('0'))
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           { to: alice.address, value: Value.fromEther('1') },
@@ -289,10 +278,10 @@ describe('execute', () => {
         delegation,
       })
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [],
-        delegate: true,
+        initialize: true,
       })
 
       const alice = privateKeyToAccount(Secp256k1.randomPrivateKey())
@@ -307,7 +296,7 @@ describe('execute', () => {
       expect(balances_before[1]).toEqual(Value.fromEther('0'))
       expect(balances_before[2]).toEqual(Value.fromEther('0'))
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         account,
         calls: [
           { to: alice.address, value: Value.fromEther('1') },
@@ -331,6 +320,46 @@ describe('execute', () => {
   })
 })
 
+describe('from', () => {
+  test('default', () => {
+    const account = DelegatedAccount.from({
+      address: '0x0000000000000000000000000000000000000001',
+      delegation: '0x0000000000000000000000000000000000000002',
+      keys: [
+        Key.fromP256({
+          expiry: 42069,
+          privateKey:
+            '0x59ff6b8de3b3b39e94b6f9fc0590cf4e3eaa9b6736e6a49c9a6b324c4f58cb9f',
+          role: 'admin',
+        }),
+      ],
+      label: 'test',
+    })
+
+    expect(account).toMatchInlineSnapshot(`
+      {
+        "address": "0x0000000000000000000000000000000000000001",
+        "delegation": "0x0000000000000000000000000000000000000002",
+        "keys": [
+          {
+            "expiry": 42069,
+            "publicKey": {
+              "prefix": 4,
+              "x": 106772332543853129808885768556237579026047721040871197501406793325324706664735n,
+              "y": 110339676533900797749866890631856969312064778492080444630962038345769628880904n,
+            },
+            "role": "admin",
+            "sign": [Function],
+            "type": "p256",
+          },
+        ],
+        "label": "test",
+        "type": "delegated",
+      }
+    `)
+  })
+})
+
 describe('prepareExecute', () => {
   describe('authorize', () => {
     test('delegated: false, key: owner, keysToAuthorize: [P256], executor: JSON-RPC', async () => {
@@ -342,7 +371,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await Delegation.prepareExecute(
+      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
         client,
         {
           account,
@@ -351,7 +380,7 @@ describe('prepareExecute', () => {
               key: keyToAuthorize,
             }),
           ],
-          delegate: true,
+          initialize: true,
         },
       )
 
@@ -359,13 +388,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -376,19 +405,14 @@ describe('prepareExecute', () => {
     })
 
     test('delegated: true, key: owner, keysToAuthorize: [P256], executor: JSON-RPC', async () => {
-      const { account, privateKey } = await getAccount(client, {
+      const { account } = await getAccount(client, {
         delegation,
       })
 
-      const authorization = await signAuthorization(client, {
-        account: privateKeyToAccount(privateKey),
-        contractAddress: delegation,
-        delegate: true,
-      })
-      await sendTransaction(client, {
-        authorizationList: [authorization],
-        account: null,
-        to: account.address,
+      await DelegatedAccount.execute(client, {
+        account,
+        calls: [],
+        initialize: true,
       })
 
       const key = account.keys[0]
@@ -397,7 +421,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await Delegation.prepareExecute(
+      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
         client,
         {
           account,
@@ -413,13 +437,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -440,7 +464,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await Delegation.prepareExecute(
+      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
         client,
         {
           account,
@@ -449,7 +473,7 @@ describe('prepareExecute', () => {
               key: keyToAuthorize,
             }),
           ],
-          delegate: true,
+          initialize: true,
           executor: privateKeyToAccount(privateKey),
         },
       )
@@ -458,13 +482,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await Delegation.execute(client, {
+      await DelegatedAccount.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await Delegation.keyAt(client, {
+        await DelegatedAccount.keyAt(client, {
           account,
           index: 0,
         }),
@@ -478,26 +502,21 @@ describe('prepareExecute', () => {
 
 describe('getExecuteSignPayload', () => {
   test('default', async () => {
-    const { account, privateKey } = await getAccount(client, {
+    const { account } = await getAccount(client, {
       delegation,
     })
 
-    const authorization = await signAuthorization(client, {
-      account: privateKeyToAccount(privateKey),
-      contractAddress: delegation,
-      delegate: true,
-    })
-    await sendTransaction(client, {
-      authorizationList: [authorization],
-      account: null,
-      to: account.address,
+    await DelegatedAccount.execute(client, {
+      account,
+      calls: [],
+      initialize: true,
     })
 
     const key = Key.createP256({
       role: 'admin',
     })
 
-    const payload = await Delegation.getExecuteSignPayload(client, {
+    const payload = await DelegatedAccount.getExecuteSignPayload(client, {
       account,
       calls: [
         Call.authorize({
@@ -517,7 +536,7 @@ describe('getExecuteSignPayload', () => {
       role: 'admin',
     })
 
-    const payload = await Delegation.getExecuteSignPayload(client, {
+    const payload = await DelegatedAccount.getExecuteSignPayload(client, {
       account,
       calls: [
         Call.authorize({
